requirements.txt
--find-links https://download.pytorch.org/whl/cu116/torch_stable
# cmake
 #wheel
# setuptools==57.5.0
# pyhash==0.9.3
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib~=3.6.3
opencv-python~=4.10.0.84
omegaconf~=2.1.2
plotly
ftfy~=6.2.3
torch==2.0.1
torchsde~=0.2.6
torchdiffeq~=0.2.4
transformers==4.25.1
vit-pytorch==1.4.4
pytorch-lightning==1.9.5
voltron-robotics
gitpython~=3.1.43
scipy~=1.9.3
sentence-transformers
torchsde
einops~=0.8.0
torchdiffeq
gym~=0.26.2
moviepy~=1.0.3
tqdm~=4.66.5
termcolor~=2.4.0
wandb~=0.17.8
einops_exts
regex~=2024.7.24
pillow~=10.4.0
torchvision~=0.15.2
pyhash~=0.9.3
litdata~=0.2.26
pybullet~=2.8.1
trimesh~=4.2.0
urdfpy~=0.0.22
pyrender~=0.1.45
tacto~=0.0.3
setuptools~=57.5.0
numpy-quaternion~=2023.0.4